backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
web3==6.20.1
httpx==0.27.2
pydantic==2.8.2